Added getAll and get requests to modules: categories, materials, patterns
