--- modules/categories/categories.services.js
@@ -1,21 +1,0 @@
-const Category = require('../../models/Category');
-
-class CategoriesService {
-  getAllCategories() {
-    return Category.find();
-  }
-
-  getCategoryById(id) {
-    return Category.findById(id);
-  }
-
-  async addCategory(data) {
-    const category = await Category(data);
-    await category.save();
-  }
-
-  deleteCategory(id) {
-    return Category.findByIdAndDelete(id);
-  }
-}
-module.exports = new CategoriesService();
