+++ modules/currencies/currencies.model.js
@@ -1,0 +1,7 @@
+const currencyType = `    
+    _id: ID!
+    date: String!
+    convertOptions: [ConvertOption!]
+`;
+
+module.exports = currencyType;
