+++ modules/currencies/currencies.resolver.js
@@ -1,0 +1,8 @@
+const Currency = require('./currencies.service');
+
+const currencyQuery = {
+  getAllCurrencies: () => Currency.getAllCurrencies(),
+  getCurrencyById: (parent, args) => Currency.getCurrencyById(args.id),
+};
+
+module.exports = currencyQuery;
