+++ modules/patterns/patterns.resolver.js
@@ -1,0 +1,8 @@
+const Pattern = require('./patterns.service');
+
+const patternQuery = {
+  getAllPatterns: () => Pattern.getAllPatterns(),
+  getPatternById: (parent, args) => Pattern.getPatternById(args.id),
+};
+
+module.exports = patternQuery;
