--- modules/patterns/patterns.services.js
@@ -1,22 +1,0 @@
-const Pattern = require('../../models/Pattern');
-
-class PatternsService {
-  getAllPatterns() {
-    return Pattern.find();
-  }
-
-  getPatternById(id) {
-    console.log(id)
-    return Pattern.findById(id);
-  }
-
-  async addPattern(data) {
-    const pattern = await Pattern(data);
-    await pattern.save();
-  }
-
-  deletePattern(id) {
-    return Pattern.findByIdAndDelete(id);
-  }
-}
-module.exports = new PatternsService();
